stock APIs added

+++ utils/SchemaValidations/stockvalidation.js
@@ -4,9 +4,9 @@ const stockRecordSchema = Joi.object({
     uid: Joi.number().integer().required(),
     stockSymbol: Joi.string().required(),
     stockName: Joi.string().required(),
-    perSharePrice: Joi.number().float().required(),
+    perSharePrice: Joi.number().required(),
     numberOfShares: Joi.number().integer().required(),
-    totalAmount: Joi.number().float().required()
+    totalAmount: Joi.number().required()
 })
 
 module.exports = {
